+++ docs/changelog.md
@@ -89,6 +89,11 @@
 
 ## 更新说明详情
 
+1.18.7 fix: fix stock_us_valuation_baidu interface
+
+	1. 修复 stock_us_valuation_baidu 接口
+	2. 修复 tool_trade_date_hist_sina 接口
+
 1.18.6 fix: fix fund_fee_em interface
 
 	1. 修复 fund_fee_em 接口
@@ -6047,6 +6052,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.7 fix: fix stock_us_valuation_baidu interface
 
 1.18.6 fix: fix fund_fee_em interface